refactor: Überarbeitung der Kalender-Filterung und UI-Komponenten
Optimierung der Konsistenz und Performance der Kalenderansichten. Die Filterlogik
wurde zentralisiert und die Benutzeroberfläche für eine bessere Benutzerführung
vereinheitlicht.

UI & Design:
- Farbschemata für Kalenderereignisse und Filter-Tags systemweit vereinheitlicht.
- Layout der Suchergebnisse für eine bessere Übersichtlichkeit optimiert.
- UI-Komponenten überarbeitet, um visuelle Diskrepanzen zwischen verschiedenen Kalenderseiten zu beheben.

Suche & Filtermanagement:
- Suchfunktionalität durch effizienteres Handling von Benutzereingaben verbessert.
- Filterverwaltung zentralisiert, um redundante State-Updates zu vermeiden.
- Interaktionslogik für Suchergebnisse verfeinert (schnellere Reaktionszeiten).

Bereinigung & Optimierung:
- Nicht verwendete Dateien und obsolete Utility-Funktionen entfernt.
- Datenfluss zwischen Filter-Sidebar und Kalender-Grid gestrafft.
- Unnötige Re-Renders durch Optimierung der Filter-Hooks reduziert.
