+++ lib/features/calendar/presentation/pages/calendar_search_page.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
 import 'package:scrollable_positioned_list/scrollable_positioned_list.dart';
@@ -7,7 +9,7 @@ import '../../domain/models/calendar_entry.dart';
 import '../providers/calendar_providers.dart';
 import '../theme/calendar_presentation_theme.dart';
 import '../widgets/event_list/cards/calendar_entry_card.dart';
-import 'search_results/search_results_sections.dart';
+import '../widgets/search_results/search_results_sections.dart';
 
 class CalendarSearchPage extends ConsumerStatefulWidget {
   const CalendarSearchPage({required this.query, super.key});
@@ -26,6 +28,7 @@ class _CalendarSearchPageState extends ConsumerState<CalendarSearchPage> {
   double _stickyHeaderPushOffset = 0;
   double _listViewportHeight = 0;
   List<_SearchListRow> _rowsForSticky = const <_SearchListRow>[];
+  bool _hasUserInteractedWithList = false;
 
   @override
   void initState() {
@@ -39,6 +42,7 @@ class _CalendarSearchPageState extends ConsumerState<CalendarSearchPage> {
     if (oldWidget.query != widget.query) {
       _stickyDay = null;
       _stickyHeaderPushOffset = 0;
+      _hasUserInteractedWithList = false;
     }
   }
 
@@ -98,22 +102,35 @@ class _CalendarSearchPageState extends ConsumerState<CalendarSearchPage> {
             LayoutBuilder(
               builder: (context, constraints) {
                 _listViewportHeight = constraints.maxHeight;
-                return ScrollablePositionedList.builder(
-                  key: ValueKey('search-list-${widget.query}'),
-                  itemPositionsListener: _itemPositionsListener,
-                  padding: const EdgeInsets.only(top: _stickyHeaderHeight),
-                  itemCount: rows.length,
-                  initialScrollIndex: targetRowIndex,
-                  itemBuilder: (context, index) {
-                    final row = rows[index];
-                    if (row.headerDay != null) {
-                      return _buildDayHeader(context: context, day: row.headerDay!);
+                return NotificationListener<ScrollNotification>(
+                  onNotification: (notification) {
+                    if (_hasUserInteractedWithList) return false;
+                    if (notification is ScrollUpdateNotification &&
+                        notification.dragDetails != null) {
+                      _hasUserInteractedWithList = true;
+                    } else if (notification is UserScrollNotification &&
+                        notification.direction != ScrollDirection.idle) {
+                      _hasUserInteractedWithList = true;
                     }
-                    return CalendarEntryCard(
-                      entry: row.entry!,
-                      applyPastStyling: true,
-                    );
+                    return false;
                   },
+                  child: ScrollablePositionedList.builder(
+                    key: ValueKey('search-list-${widget.query}'),
+                    itemPositionsListener: _itemPositionsListener,
+                    padding: const EdgeInsets.only(top: _stickyHeaderHeight),
+                    itemCount: rows.length,
+                    initialScrollIndex: targetRowIndex,
+                    itemBuilder: (context, index) {
+                      final row = rows[index];
+                      if (row.headerDay != null) {
+                        return _buildDayHeader(context: context, day: row.headerDay!);
+                      }
+                      return CalendarEntryCard(
+                        entry: row.entry!,
+                        applyPastStyling: true,
+                      );
+                    },
+                  ),
                 );
               },
             ),
@@ -244,9 +261,12 @@ class _CalendarSearchPageState extends ConsumerState<CalendarSearchPage> {
       }
     }
 
-    if (_stickyDay == null ||
-        !_isSameDay(_stickyDay!, nextStickyDay) ||
-        (_stickyHeaderPushOffset - pushOffset).abs() > 0.1) {
+    final stickyDayChanged = _stickyDay == null || !_isSameDay(_stickyDay!, nextStickyDay);
+    if (stickyDayChanged && _hasUserInteractedWithList) {
+      HapticFeedback.mediumImpact();
+    }
+
+    if (stickyDayChanged || (_stickyHeaderPushOffset - pushOffset).abs() > 0.1) {
       setState(() {
         _stickyDay = nextStickyDay;
         _stickyHeaderPushOffset = pushOffset;
