+++ lib/features/calendar/presentation/providers/filter/calendar/calendar_filter_options_providers.dart
@@ -1,14 +1,14 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart' as fr;
 
 import '../../../../../../core/database/backend_enums.dart';
+import '../../../../domain/filter/calendar_filter_text.dart';
 import '../../../../domain/models/calendar_entry.dart';
-import 'calendar_filter_utils.dart';
 import '../../calendar_providers.dart';
 
 final calendarChoirFilterOptionsProvider = fr.Provider<List<String>>((ref) {
   return BackendChoir.values
       .where((value) => value != BackendChoir.unknown)
-      .map((value) => normalizeFilterText(value.toBackend()))
+      .map((value) => normalizeCalendarFilterText(value.toBackend()))
       .whereType<String>()
       .toList(growable: false);
 });
@@ -16,7 +16,7 @@ final calendarChoirFilterOptionsProvider = fr.Provider<List<String>>((ref) {
 final calendarVoiceFilterOptionsProvider = fr.Provider<List<String>>((ref) {
   return BackendVoice.values
       .where((value) => value != BackendVoice.unknown)
-      .map((value) => normalizeFilterText(value.toBackend()))
+      .map((value) => normalizeCalendarFilterText(value.toBackend()))
       .whereType<String>()
       .toList(growable: false);
 });
@@ -35,7 +35,7 @@ final calendarClassFilterOptionsProvider =
     });
 
 void _addClassName(Set<String> set, CalendarEntry entry) {
-  final className = normalizeFilterText(entry.className);
+  final className = normalizeCalendarFilterText(entry.className);
   if (className != null) {
     set.add(className);
   }
