+++ lib/features/calendar/presentation/widgets/search_results/search_results_sections.dart
